pyproject.toml
[project]
name = "shl-recommender-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
